Delete post regardless of amount of votes on post

+++ src/resolvers/post.ts
@@ -127,9 +127,40 @@ export class PostResolver {
   @UseMiddleware(isAuth)
   async deletePost(
     @Arg('id', () => Int) id: number,
-    @Ctx() { prisma }: myContext
+    @Ctx() { req, prisma }: myContext
   ): Promise<Boolean> {
-    await prisma.post.delete({ where: { id } });
-    return true;
-  }
+    console.log('starting delete from req.session:', req.session);
+
+    const post = await prisma.post.findUnique({ where: { id } });
+
+    console.log(post);
+
+    console.log('starting checks');
+
+    if (!post) {
+      console.log('no matching Id');
+      return false;
+    }
+    
+    if (post.authorId !== req.session.userId) {
+      console.log('you can do that!');
+      throw new Error('Not Authorized');
+    }
+
+    if (post.authorId === req.session.userId) {
+      console.log('req.session.user matches post.authorId');
+    }
+
+    // we have to delete votes on a post before deleting the post
+    // alternatively could use onDelete: CASCADE in schema.prisma
+    await prisma.votes.deleteMany({ where: { postId: id } });
+
+    try {
+      await prisma.post.delete({ where: { id } });
+      console.log('delete successful');
+      return true;
+    } catch (error) {
+      console.error('delete failed:', error);
+      return false;
+    }
 }
